Remove unused imports
