+++ AcceptanceTest/HomesEngland/ConsoleImporter/ConsoleImporterAcceptanceTests.cs
@@ -1,6 +1,4 @@
-﻿using System;
-using System.IO;
-using System.Reflection;
+﻿using System.IO;
 using System.Threading.Tasks;
 using System.Transactions;
 using FluentAssertions;
