+++ AcceptanceTest/HomesEngland/UseCases/GenerateAssetsUseCaseAcceptanceTests.cs
@@ -5,7 +5,6 @@ using System.Transactions;
 using FluentAssertions;
 using HomesEngland.UseCase.GenerateAssets;
 using HomesEngland.UseCase.GenerateAssets.Models;
-using HomesEngland.UseCase.GetAsset;
 using HomesEngland.UseCase.GetAsset.Models;
 using HomesEngland.UseCase.SearchAsset;
 using HomesEngland.UseCase.SearchAsset.Models;
