+++ AcceptanceTest/HomesEngland/UseCases/Search/SearchUseCaseAcceptanceTests.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using System.Runtime.InteropServices;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Transactions;
