+++ HomesEngland.Gateway.DataGenerator/Program.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using HomesEngland.UseCase.GenerateAssets;
+﻿using HomesEngland.UseCase.GenerateAssets;
 using Main;
 
 namespace HomesEngland.Gateway.DataGenerator
