+++ HomesEngland.Gateway/Sql/Postgres/IDatabaseConnectionStringFormatter.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace HomesEngland.Gateway
 {
     public interface IDatabaseConnectionStringFormatter
