+++ HomesEnglandTest/UseCase/ImportAssets/ImportAssetsUseCaseTest.cs
@@ -10,10 +10,7 @@ using HomesEngland.UseCase.GetAsset.Models;
 using HomesEngland.UseCase.ImportAssets;
 using HomesEngland.UseCase.ImportAssets.Impl;
 using HomesEngland.UseCase.ImportAssets.Models;
-using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
-using Microsoft.Extensions.Logging;
 using Moq;
-using NSubstitute.Exceptions;
 using NUnit.Framework;
 
 namespace HomesEnglandTest.UseCase.ImportAssets
