+++ WebApi/Extensions/ResponseData.cs
@@ -1,0 +1,12 @@
+namespace WebApi.Extensions
+{
+    public class ResponseData<T>
+    {
+        public T Data { get; }
+
+        public ResponseData(T data)
+        {
+            Data = data;
+        }
+    }
+}
