+++ WebApi/Settings/AppSettings.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace WebApi.Settings
+﻿namespace WebApi.Settings
 {
     public class AppSettings
     {
