+++ WebApiTest/Controller/Asset/CalculateAssetAggregates/CalculateAssetAggregatesControllerTests.cs
@@ -5,10 +5,8 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using HomesEngland.UseCase.CalculateAssetAggregates;
 using HomesEngland.UseCase.CalculateAssetAggregates.Models;
-using Infrastructure.Api.Response;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Server.Kestrel.Core.Features;
 using Microsoft.Extensions.Primitives;
 using Moq;
 using NUnit.Framework;
